Trying to get a build server

+++ Info Board/src/me/xxsniperzzxxsd/infoboard/Variables/VaultVariables.java
@@ -13,7 +13,7 @@ public class VaultVariables {
 		if (InfoBoard.economy != null && InfoBoard.economyB)
 		{
 			if (newString.contains("<money>"))
-				newString = newString.replaceAll("<money>", String.valueOf((int) InfoBoard.economy.getBalance(player.getName())));
+				newString = newString.replaceAll("<money>", String.valueOf((long)InfoBoard.economy.getBalance(player.getName())));
 
 			if (newString.contains("<vaultcurrencyplural>"))
 				newString = newString.replaceAll("<vaultcurrencyplural>", String.valueOf(InfoBoard.economy.currencyNamePlural()));
